chore: move pool simulator types to a single file
