+++ apps/balancer-tools/src/app/poolsimulator/(components)/SearchPoolFormDialog.tsx
@@ -4,11 +4,9 @@ import { PropsWithChildren } from "react";
 
 import { Dialog } from "#/components/Dialog";
 import { SearchPoolForm } from "#/components/SearchPoolForm";
-import {
-  PoolType,
-  PoolTypeEnum,
-  usePoolSimulator,
-} from "#/contexts/PoolSimulatorContext";
+import { PoolType, usePoolSimulator } from "#/contexts/PoolSimulatorContext";
+
+import { PoolTypeEnum } from "../(types)";
 
 const poolTypes = {
   [PoolTypeEnum.MetaStable]: ["Stable", "MetaStable", "ComposableStable"],
