+++ apps/balancer-tools/src/contexts/PoolSimulatorContext.tsx
@@ -13,47 +13,17 @@ import {
 } from "react";
 
 import {
+  GyroEParams,
+  MetaStableParams,
+  PoolTypeEnum,
+  TokensData,
+} from "#/app/poolsimulator/(types)";
+import {
   convertAnalysisDataToAMM,
   convertGqlToAnalysisData,
 } from "#/app/poolsimulator/(utils)";
 import { PoolAttribute } from "#/components/SearchPoolForm";
 import { pools } from "#/lib/gql";
-
-export interface TokensData {
-  symbol: string;
-  balance: number;
-  decimal: number;
-  rate?: number;
-  weight?: number;
-}
-
-export interface MetaStableParams {
-  ampFactor?: number;
-  swapFee?: number;
-}
-
-export interface GyroEParams {
-  alpha?: number;
-  beta?: number;
-  lambda?: number;
-  c?: number;
-  s?: number;
-  swapFee?: number;
-  tauAlphaX?: number;
-  tauAlphaY?: number;
-  tauBetaX?: number;
-  tauBetaY?: number;
-  u?: number;
-  v?: number;
-  w?: number;
-  z?: number;
-  dSq?: number;
-}
-
-export enum PoolTypeEnum {
-  MetaStable = "MetaStable",
-  GyroE = "GyroE",
-}
 
 export type PoolParams = MetaStableParams & GyroEParams;
 export type PoolType = PoolTypeEnum;
